lxd/storage/filesystem: Adds filesystem package and moves FilesystemDetect to it as just Detect
