+++ lxd/storage/filesystem/fs_32bit.go
@@ -1,7 +1,7 @@
 //go:build 386 || arm || ppc || s390 || mips || mipsle
 // +build 386 arm ppc s390 mips mipsle
 
-package util
+package filesystem
 
 const (
 	// FilesystemSuperMagicBtrfs is the 32bit magic for Btrfs (as signed constant)
