+++ lxd/storage/filesystem/fs_64bit.go
@@ -1,7 +1,7 @@
 //go:build amd64 || ppc64 || ppc64le || arm64 || s390x || mips64 || mips64le || riscv64
 // +build amd64 ppc64 ppc64le arm64 s390x mips64 mips64le riscv64
 
-package util
+package filesystem
 
 const (
 	// FilesystemSuperMagicBtrfs is the 64bit magic for Btrfs
